Add AWS Certification Information.
As of January 5th 2017, I am (finally) an AWS Certified Solutions Architect, Associate level.  I'll be writing a blog post about my experience preparing for and taking the exam shortly.  If you'll be taking the exam soon, be sure to subsribe to the RSS feed (if you're reading this in early January) or check out the blog post (if this is the magical future).